'recent' state indicator on lists, but also refactoring episode api request into separate util file and applying to all places where the state indicator is needed
