+++ app/src/components/PodcastEpisodesView.js
@@ -11,6 +11,7 @@ import { connect } from 'react-redux';
 import fetch from '../util/fetch';
 import moment from 'moment';
 import { getPinnedEpisodes } from '../util/pins';
+import { getEpisodesFeed } from '../util/feeds';
 import Loader from './Loader';
 
 class PodcastEpisodesView extends React.Component {
@@ -19,7 +20,7 @@ class PodcastEpisodesView extends React.Component {
 		this.toggleFollowPodcast = this.toggleFollowPodcast.bind(this);
 		this.toggleMenu = this.toggleMenu.bind(this);
 		this.state = {
-			episodeCursor: 1,
+			episodeCursor: 1, // mongoose-api-query starts pages at 1, not 0
 			menuIsOpen: false,
 			sortBy: 'latest',
 		};
@@ -28,14 +29,16 @@ class PodcastEpisodesView extends React.Component {
 		this.props.getPodcast(this.props.match.params.podcastID);
 		this.getEpisodes(this.props.match.params.podcastID);
 		getPinnedEpisodes(this.props.dispatch);
+		getEpisodesFeed(this.props.dispatch, 0, 20); // this is to populate 'recent' state indicators
 	}
 	componentWillReceiveProps(nextProps) {
-		if (nextProps.match.params.podcastID !== this.props.match.params.podcastID) {
+		if (nextProps.match.params.podcastID !== this.props.match.params.podcastID) { // essentially, if we've just switched views from one podcast to another
 			this.props.getPodcast(nextProps.match.params.podcastID);
 			this.getEpisodes(nextProps.match.params.podcastID);
 			getPinnedEpisodes(this.props.dispatch);
+			getEpisodesFeed(this.props.dispatch, 0, 20); // this is to populate 'recent' state indicators
 			this.setState({
-				episodeCursor: 1,
+				episodeCursor: 1, // mongoose-api-query starts pages at 1, not 0
 			});
 		}
 	}
@@ -315,6 +318,22 @@ const mapStateToProps = (state, ownProps) => {
 		} else {
 			episode.pinned = false;
 		}
+
+		if (state.feeds && state.feeds[`user_episode:${localStorage['authedUser']}`]) {
+			if (
+				state.feeds[`user_episode:${localStorage['authedUser']}`].indexOf(
+					episode._id,
+				) < 20 &&
+				state.feeds[`user_episode:${localStorage['authedUser']}`].indexOf(
+					episode._id,
+				) !== -1
+			) {
+				episode.recent = true;
+			} else {
+				episode.recent = false;
+			}
+		}
+
 		// attach podcast
 		episode.podcast = { ...podcast };
 	}
